log4js: remove the default file appender to avoid startup problems
The file appender was configured to put the logs into `log/demo.log`,
which depends on the `log` directory existing. However, this path is
relative to the cwd, not the location where outrigger is installed,
so it fails when not running from the source tree or if ./log doesn't
happen to exist.

As a short-term fix, just remove the default log appender entry. Users
that want a file log in addition to console can always configure
log4js manually.