Add docker monitoring examples using cadvisor and influx.
examples/cadvisor-influx/dc-cadvisor-influx.yml is a docker-compose
file that when combined with dc-outrigger.yml results in a properly
linked cadvisor, influxdb, and outrigger.

examples/cadvisor-influx/README.md includes notes on config, setup,
and a brief overview of the programs.

Add necessary config to examples/juttle-config.json to refer to the
services from the docker-compose setup.

Add entries to examples/index.juttle to refer to cadvisor-influx
directory.

Also ran into a problem with specifying more than one override .yml
file during docker-compose up. It turns out that the links directive
is not merged between multiple files, but overwritten for each
additional file.

This meant that a links: entry in dc-elastic.yml was being overwritten
with the links: entry in dc-cadvisor-influx.yml, resulting in
outrigger not being able to reach elasticsearch.

The fix for this is to use external_links, which *is* merged across
multiple .yml files. When using external_links, however, you have to
refer to explicit container names instead of services from the .yml
file, so add some explicit names for the elasticsearch and influxdb
containers, and use those names in the juttle-config files.

dc-elastic/dc-cadvisor-influx keep the links: entries, but that's only
used to make sure the containers are started up in the right
order. the external_links entry is what actually ensures the
containers can communicate with each other.