Get standalone docker image containing core demo programs.
Update the outrigger docker image to not explicitly install the
adapters or juttle any longer. This is no longer required with #51.

Add a new juttle program index.juttle that creates a table containing
markdown links to all the core-demo juttle programs.

The base outrigger Dockerfile now copies all example juttle programs
to /opt/outrigger/juttles/examples. Outriggerd runs with a --root of
/opt/outrigger/juttles, so the examples are accessable via urls like
run?path=/examples/core-juttle/hello_world.juttle.

The remaining demos should be able to use this base image with
additional options to mount in other juttle programs and/or config
files as necessary.

Add a build.sh script that handles the details of copying the juttle
scripts to a directory below the current directory (relative parent
directories aren't allowed in COPY commands) and running the docker
build command.

Modify the README.md to change the install instructions to refer to
the juttle/outrigger image and stop using ``outrigger-client
browser``.