Get docker compose based setup working.
This updates the demo instructions to use docker compose instead of
individual steps to run the individual containers, populate them with
the news data, etc. It generally results in a one-step setup, which is
much simpler.

The docker compose config contains 3 containers:
 - elasticsearch
 - logstash
 - outrigger

Elasticsearch and logstash are linked, and outrigger and elasticsearch
are linked. The linking is handled by docker-compose, so you don't
need to specify any custom hostnames as a part of setup any longer.

Logstash maps in logstash.conf and the set of news data. The command
for the logstash container is overridden to run "unzip news_data.zip |
logstash", so it automatically parses the data and populates
elasticsearch on startup.

Outrigger maps the config from juttle-config.json to
/opt/outrigger/.juttle-config.json and maps the juttles directory to
/opt/outrigger/juttles.

The logstash and juttle configs use hostnames from the docker-compose
file.

Update the README to use the much simpler set of instructions, discuss
running the demo when not using docker-machine, and fix the links to
the juttle programs from the readme. Also the instructions generally
guide people towards clicking on links from index.juttle rather than
running docker-exec or the juttle cli (which isn't directly available
in the current outrigger).

Make a couple of minor changes to the juttle programs themselves--add
a description column to the table, fix some incorrect tags in the
rollup programs, add a program to show the number of elastic points,
add a link to search_ui.juttle.

I'm going to make additional changes based on the outrigger docker
image that I'm working on in
https://github.com/juttle/outrigger/pull/54, but I'll do that in a
later commit once that is merged.