Update to run via docker.
Update the instructions to use docker and the outrigger image.

Move the juttle program to ./juttles for consistency with the other demos.

Update the instructions to use a juttle-config.json file in the current directory for consistency with other demos.

Update the instructions to visit the run url to view the program instead of using outrigger-client browser.