Address more PR feedback from dave.
Just return the promises instead of calling done() when possible.

In #73 dave set the timeout to 30 seconds via gulp, so remove places
where I was setting a timeout <= 30 seconds.

Also I noticed another problem while running gulp tests. When a
delayed shutdown was happening, the websocket wasn't being torn down
immediately. It would *eventually* be shut down, but only due to a
lack of ping messages. This was for 2 reasons:

 - I wasn't using the right call to check the queue being empty,
   should use isEmpty().

 - I need to perform the check and call to close() at the end of the
   drain() path, as drain() won't be called when the queue is empty.

So fix those problems.