Add gmail adapter examples.
Add a note to the top level README/index.juttle referring to the gmail directory.

In the gmail directory, add a index.juttle, a README.md, and 3 programs:

 - messages_by_sender: read the last 12 hours of messages and categorize by sender. You can specify any gmail search expression in an input.
 - messages_by_time: read the last 12 hours of messages sent to "me" and categorize by time received.
 - write_results: write some artificial data to gmail (i.e. send in an email).

Update outrigger to v0.3.0 of the gmail adapter so it has write support and support for filter expressions.

In order for these examples to work with the outrigger docker image,
I'll have to npm publish outrigger 0.4.0 and publish a new docker
image to pick up the juttle-gmail-adapter version. I'll do that
shortly after merging if not tomorrow.