Create a docker image containing the preloaded news data.
To even further reduce the startup time, docker-compose and the
elasticsearch container support volumes, which are sets of data that
live alongside containers that are created when the container is
created. This gives us an easy way to provide a preloaded
elasticsearch data directory--by providing a image that has the
contents of /usr/share/elasticsearch/data from a prior full parsing +
loading instance.

Create an image juttle/elastic_news_data, which is an image containing
the /usr/share/elasticsearch/data directory created from a previously
built docker-compose environment. See the README.md in
docker/images/elastic_news_data for the full set of steps to create
the container).