Split JuttleJob into a base and child class.
JuttleJob is now solely responsible for managing the subprocess,
emitting events when the subprocess closes, etc.

It has two methods that should be overridden by derived classes:
  - send_to_clients: send a message to any clients related to this job.
  - close: close all connections to clients.

A new class WebsocketJuttleJob derives from JuttleJob and handles the
details of passing messages to websocket endpoints, the circular
buffer, etc.

This split made the responsibilities a bit clearer, so that also
eliminated the need for methods like job_started/job_stopped.

The Job Manager creates a WebsocketJuttleJob instead of a JuttleJob.