Add support for one-shot aka immediate program execution.
New class ImmediateJuttleJob inherits from JuttleJob and queues up the
set of points, errors, warnings, etc in memory. It has a waitfor()
method that returns a promise that resolves when the program is
complete with the program's output, or rejects when the program times
out.

While adding this class, I realized there wasn't really any need for
an explicit close() function--the object could just wait for an 'end'
event to close connections. So remove close().

Modify job-manager's run_program function to take an option
wait_for_end. If false, it creates a WebsocketJuttleJob instance,
otherwise it creates an ImmediateJuttleJob instance. When wait_for_end
is true, it calls the ImmediateJuttleJob's waitfor() method in the
promise chain that is returned. If the promise chain rejects with a
Timeout Error, call job-manager::delete_job to stop the program.

job-handlers' create_job is now split into two functions
create_job_from_bundle and create_job_from_path, one of which is
called based on whether or not the body contains a path
property. create_job_from_path has a promise chain that involves
bundling and calling job-manager's run_program with wait_for_end =
true.

Move some of the JuttleBundler configuration that specified default
paths for program module resolution into the class itself, so you get
consistent behavior from the several places where the bundler is used.

Add a new error JS-TIMEOUT-ERROR that is returned when you run a
program and the timeout kicks in.

Add docs on providing a path to the /jobs endpoint. I also noticed
that the table of errors was out of date, so added any missing errors.

Add unit tests for the following cases:
 - Negative:
    - A path that doesn't exist
    - A program with compile errors
    - A program where the timeout triggers
 - Positive:
    - Simple program with medium-complexity output
    - Program with modules
    - Program with inputs
    - Program with runtime errors
    - Program with runtime warnings

Add several example programs for all the tests.