jsdp-value-converter: add hack to handle infinite dates
The special moment values :beginning: and :end: are represented
internally as -Infinity and Infinity. While moment can (sort-of)
handle these, Date cannot, so outrigger blows up if one of these
values is used in the time range for read or in the points.

To work around this problem, convert infinite moments to the
smallest / largest legal Javascript date before encoding over
JSDP.